api_yamdb/requirements.txt
django~=2.2
django-filter~=2.1
djangorestframework~=3.12
djangorestframework-simplejwt~=4.7
gunicorn~=20.0
flake8~=5.0
flake8-broken-line~=0.6
flake8-isort~=5.0
flake8-return~=1.2
pep8-naming~=0.13
psycopg2-binary==2.8.6
PyJWT~=2.1
pytest~=6.2
pytest-django~=4.4
pytest-pythonpath~=0.7
python-dotenv~=0.19
requests~=2.26